+++ src/terra-dev-site/test/workspace/Tab6.jsx
@@ -6,6 +6,7 @@ import {
 const Tab6 = () => (
   <WorkspaceContent
     activityOverlay={<WorkspaceContentActivityOverlay variant="loading" />}
+    label="Tab 6 with Jam and Bread"
   />
 );